Validate admin/operator names in board #edit's [P] option
Relocated player_exists()/known_player_names()/is_online()/
find_players()/online_player_names() from commands/messaging.py to
tada_utilities.py (Ryan's call) -- general 'does this player name
exist/who's online' lookups with callers well outside messaging
(bar/zelda.py, bar/blue_djinn.py, and now commands/board/edit.py).
commands/messaging.py's own prompt_player_choice() re-imports
find_players()/online_player_names() for its numbered-list UI; all
five external call sites (bar/zelda.py, bar/blue_djinn.py, commands/
groups.py, commands/mail.py, commands/page.py) updated to import from
tada_utilities directly instead of via commands.messaging.

New _split_valid_names() in commands/board/edit.py, used by both
_manage_admins() (per-board) and _manage_sig_admins() (per-SIG) 'A'
(add) actions: rejects any name that isn't an online player or a saved
character, reporting 'No such player: <name>.' rather than silently
storing a typo that would never actually grant anyone admin/operator
status.

Tests: existing tests/social/test_messaging.py suite unchanged in
substance (import path only); tests/social/test_board_edit.py's
existing admin add/remove test now mocks player_exists (it was testing
UI flow, not real validation); two new tests cover the reject-unknown-
name path for both the per-board and per-SIG admin lists.
