+++ server/commands/page.py
@@ -30,8 +30,9 @@ from __future__ import annotations
 import mail as mail_store
 from commands.base_command import Command, CommandResult, Mode
 from commands.help import Help, HelpCategory
-from commands.messaging import parse_targets, expand_groups, find_online, player_exists, is_in_combat
+from commands.messaging import parse_targets, expand_groups, find_online, is_in_combat
 from network_context import GameContext
+from tada_utilities import player_exists
 
 _CONTROL_WORDS = {'#ignore', '#unignore', '#haven', '#unhaven'}
