+++ server/tada_utilities.py
@@ -1,6 +1,7 @@
 #!/bin/env python3
 
 import doctest
+import fnmatch
 import logging
 import random
 import re
@@ -669,6 +670,99 @@ async def fileread(ctx: 'GameContext', filename: str) -> None:
         await ctx.send(*lines)
     except FileNotFoundError:
         await ctx.send(f'File not found: {file_path}')
+
+
+# ---------------------------------------------------------------------------
+# Player lookup by name
+# ---------------------------------------------------------------------------
+# Moved here from commands/messaging.py (Ryan's call, 2026-09-01) -- these
+# are general "does this player name exist/who's online" checks with no
+# messaging-specific logic, and were already reached for from well outside
+# messaging (bar/zelda.py, bar/blue_djinn.py, commands/board/edit.py's
+# admin-list validation). commands/messaging.py's own prompt_player_choice()
+# still imports find_players()/online_player_names() from here for its
+# numbered-list UI.
+
+def online_player_names(server) -> list[str]:
+    """Return display names of all currently connected players."""
+    names = []
+    for client in server.clients.values():
+        ctx  = getattr(client, 'ctx', None)
+        name = getattr(getattr(ctx, 'player', None), 'name', '')
+        if name:
+            names.append(name)
+    return names
+
+
+def known_player_names() -> list[str]:
+    """Return ids of all players that have a save file, online or not.
+
+    Names are derived from the filename: run/server/player-<name>.json.
+    """
+    import glob
+    import os
+    try:
+        import net_common
+        base = getattr(net_common, 'run_server_dir', None) or './run/server'
+    except Exception:
+        base = './run/server'
+    names = []
+    for path in glob.glob(os.path.join(str(base), 'player-*.json')):
+        stem = os.path.basename(path)[len('player-'):-len('.json')]
+        if stem:
+            names.append(stem)
+    return names
+
+
+def is_online(server, name: str) -> bool:
+    """Return True if a player with this name is currently connected."""
+    needle = name.lower()
+    return any(n.lower() == needle for n in online_player_names(server))
+
+
+def find_players(server, pattern: str) -> list[str]:
+    """Return sorted names matching pattern, checking online clients then save files.
+
+    Supports shell-style wildcards: * matches any string, ? matches one character.
+    Matching is case-insensitive.  Each name appears at most once.
+
+    Examples:
+        find_players(server, '*')       → all known players
+        find_players(server, 'ral*')    → ['railbender'] (if that save file exists)
+        find_players(server, 'r?lan')   → ['Rulan'] (if online or saved)
+    """
+    pat  = pattern.lower()
+    seen: set[str] = set()
+    results: list[str] = []
+
+    for name in online_player_names(server):
+        if fnmatch.fnmatch(name.lower(), pat):
+            key = name.lower()
+            if key not in seen:
+                seen.add(key)
+                results.append(name)
+
+    for name in known_player_names():
+        if fnmatch.fnmatch(name.lower(), pat):
+            key = name.lower()
+            if key not in seen:
+                seen.add(key)
+                results.append(name)
+
+    return sorted(results, key=str.lower)
+
+
+def player_exists(server, name: str) -> bool:
+    """Return True if the name belongs to an online player or has a save file.
+
+    Supports ? and * wildcards — returns True if any player matches.
+    """
+    if '*' in name or '?' in name:
+        return bool(find_players(server, name))
+    if is_online(server, name):
+        return True
+    needle = name.lower()
+    return any(n.lower() == needle for n in known_player_names())
 
 
 # ---------------------------------------------------------------------------
